better handling of doubles between Octave and Python
passed using exact hex format instead of imprecise strings.  Some fallout from this, in particular with indexing.  This commit also wraps the places I've found in int32().  But there might be more.

+++ inst/@sym/mat_rccross_access.m
@@ -74,8 +74,8 @@ function z = mat_rccross_access(A, r, c)
           '        M[i,j] = A[rr[i],cc[j]]\n'  ...
           'return (M,)' ];
 
-  rr = num2cell(r-1);
-  cc = num2cell(c-1);
+  rr = num2cell(int32(r-1));
+  cc = num2cell(int32(c-1));
   z = python_cmd(cmd, A, rr, cc);
 
   % FIXME: here's some code could be used for slices
